Optionally print less messages at boot
The patch optionally makes grub not show any text (be fully quiet) when
timeout_style=hidden is set and the user does not interrupt the boot.

Combined with a later patch in this series which makes grub not touch
the EFI console unless it actually has some text to print, this will keep
the vendor logo which EFI put on the display in place until the kernel
touches the display. Leading to a more smooth / seamless boot experience.

At least Fedora/RHEL/CentOS and Ubuntu have been carrying patches for this
for a long time now (since 2013). There have been several attempts to
upstream these patches in the past already, which have been rejected
because not everyone likes the quiet behavior.

This patch makes the quiet behavior optional and defaults to off, so
unless grub is compiled with the new --enable-quiet-boot configure option
this patch changes nothing.

This patch is a mix of the Fedora patches for this and:
https://git.launchpad.net/~ubuntu-core-dev/grub/+git/ubuntu/tree/debian/patches/maybe_quiet.patch
https://git.launchpad.net/~ubuntu-core-dev/grub/+git/ubuntu/tree/debian/patches/gettext_quiet.patch
Specifically the configure.ac changes for making this optional come from
the Ubuntu patches, as Fedora's patches where simply unconditionally
removing all the unwanted grub_printf calls.

Cc: Colin Watson <cjwatson@ubuntu.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>