10_linux.in: restore existence check in `get_sorted_bls`
This is necessary to handle `/boot/loader/entries` not existing
at all (or possibly existing but being empty - not sure about
that case). Without this check, this function gets pretty wacky
and winds up returning the contents of the current working
directory, which of course causes whatever called it to break.

Resolves: rhbz#1836020

Signed-off-by: Adam Williamson <awilliam@redhat.com>